Move event to a submodule
